Create To open an Excel file automatically in JavaScript, you can use the window.open() method with the path to the Excel file as the parameter
This function takes a filePath parameter that specifies the path to the Excel file you want to open. It then uses the window.open() method to open the file in a new window with the '_blank' target, which will cause the file to open in the default application for Excel files on the user's device.

The function also checks if the window.open() method returned a value, which would indicate that the file was opened successfully. If the method returns null, the function logs an error message to the console.

Note that this function will only work in a web browser environment, not in a server-side Node.js environment. Additionally, the user's device must have a default application set up to handle Excel files for this function to work correctly.